Make grub-efi work for isohybrid images
By default (at least in the Debian/buster version of live-build),
grub-efi is enabled alongside syslinux for iso and isohybrid images.
Sinec our Makefile did not call `lb binary_grub-efi`, it was not a
actually installed, but this commit changes that.

This produces a bootable configuration, just without all the extra menu
options that are present in the syslinux config (e.g. installer,
languages, etc.). It also lacks the bootfrom=removable option which is
set in syslinux cfg, rather than in auto/config.